+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.6] - 2023-11-19
+
+### Added
+
+- Added Layout Direction `dir` context in the App UI `ApplicationContext`.
+- Added support of LTR and RTL layout direction in most App UI components
+
+### Changed
+
+- Previously the Panel Constructor set the default scale context to "large" on mobile platforms, now the default scale context for any platform is "medium".
+
 ## [1.0.0-pre.5] - 2023-11-15
 
 ### Changed